requirements.txt
hydra-core==1.3.2
omegaconf==2.3.0
tokenizers==0.20.3
datasets==3.2.0
scipy==1.14.1
google-cloud-storage==2.19.0
ipython==8.28.0
ipdb==0.13.13
flax==0.10.2
torchdata==0.8.0
wandb==0.19.1
pytest==8.3.5
lingvo==0.12.7
optax==0.2.4
git+https://github.com/google/CommonLoopUtils@307b0bc65ae2e2d801b6c19df2431c060a0aa4ec